pyproject.toml
[project]
name = "projeto-oficina"
version = "0.1.0"
description = ""
authors = [
    {name = "Arlindo Neto",email = "netonery06@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13.7,<4.0"
dependencies = [
    "fastapi[standard] (>=0.119.0,<0.120.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)"
]

[tool.poetry]
packages = [{include = "projeto_oficina", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.0,<0.15.0)"
]

[tool.ruff]
line-length = 100
extend-exclude = ['alembic']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'double'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'

